implement gm17, verification contract not working for now
