+++ zokrates_core/build.rs
@@ -26,7 +26,9 @@ fn main() {
             .include(libsnark_source_path.join("depends/libff"))
             .include(libsnark_source_path.join("depends/libfqfft"))
             .define("CURVE_ALT_BN128", None)
-            .file("lib/wraplibsnark.cpp")
+            .file("lib/util.cpp")
+            .file("lib/gm17.cpp")
+            .file("lib/pghr13.cpp")
             .compile("libwraplibsnark.a");
 
         cc::Build::new()
